+++ js/ui/drop_down_button.js
@@ -17,7 +17,7 @@ import { isPlainObject, isDefined } from '../core/utils/type';
 import { ensureDefined } from '../core/utils/common';
 import Guid from '../core/guid';
 import { format as formatMessage } from '../localization/message';
-import { getElementWidth, getPopupWidth } from './drop_down_editor/drop_down';
+import { getElementWidth, getPopupWidth } from './drop_down_editor/utils';
 
 const DROP_DOWN_BUTTON_CLASS = 'dx-dropdownbutton';
 const DROP_DOWN_BUTTON_CONTENT = 'dx-dropdownbutton-content';
